OSD-7348 Refactor upgraders into universal steps
