+++ pkg/cluster_upgrader_builder/cluster_upgrader_builder.go
@@ -1,21 +1,21 @@
 package cluster_upgrader_builder
 
 import (
+	"context"
 	"github.com/go-logr/logr"
 	"github.com/openshift/managed-upgrade-operator/pkg/eventmanager"
+	"github.com/openshift/managed-upgrade-operator/pkg/upgraders"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
 	upgradev1alpha1 "github.com/openshift/managed-upgrade-operator/pkg/apis/upgrade/v1alpha1"
 	"github.com/openshift/managed-upgrade-operator/pkg/configmanager"
 	"github.com/openshift/managed-upgrade-operator/pkg/metrics"
-	"github.com/openshift/managed-upgrade-operator/pkg/upgraders/aro"
-	"github.com/openshift/managed-upgrade-operator/pkg/upgraders/osd"
 )
 
 // Interface describing the functions of a cluster upgrader.
 //go:generate mockgen -destination=mocks/cluster_upgrader.go -package=mocks github.com/openshift/managed-upgrade-operator/pkg/cluster_upgrader_builder ClusterUpgrader
 type ClusterUpgrader interface {
-	UpgradeCluster(upgradeConfig *upgradev1alpha1.UpgradeConfig, logger logr.Logger) (upgradev1alpha1.UpgradePhase, *upgradev1alpha1.UpgradeCondition, error)
+	UpgradeCluster(ctx context.Context, upgradeConfig *upgradev1alpha1.UpgradeConfig, logger logr.Logger) (upgradev1alpha1.UpgradePhase, *upgradev1alpha1.UpgradeCondition, error)
 }
 
 //go:generate mockgen -destination=mocks/cluster_upgrader_builder.go -package=mocks github.com/openshift/managed-upgrade-operator/pkg/cluster_upgrader_builder ClusterUpgraderBuilder
@@ -32,23 +32,22 @@ type clusterUpgraderBuilder struct{}
 func (cub *clusterUpgraderBuilder) NewClient(c client.Client, cfm configmanager.ConfigManager, mc metrics.Metrics, nc eventmanager.EventManager, upgradeType upgradev1alpha1.UpgradeType) (ClusterUpgrader, error) {
 	switch upgradeType {
 	case upgradev1alpha1.OSD:
-		cu, err := osd.NewClient(c, cfm, mc, nc)
+		cu, err := upgraders.NewOSDUpgrader(c, cfm, mc, nc)
 		if err != nil {
 			return nil, err
 		}
 		return cu, nil
 	case upgradev1alpha1.ARO:
-		cu, err := aro.NewClient(c, cfm, mc, nc)
+		cu, err := upgraders.NewAROUpgrader(c, cfm, mc, nc)
 		if err != nil {
 			return nil, err
 		}
 		return cu, nil
 	default:
-		cu, err := osd.NewClient(c, cfm, mc, nc)
+		cu, err := upgraders.NewOSDUpgrader(c, cfm, mc, nc)
 		if err != nil {
 			return nil, err
 		}
 		return cu, nil
-
 	}
 }
