+++ pkg/cluster_upgrader_builder/mocks/cluster_upgrader.go
@@ -5,6 +5,7 @@
 package mocks
 
 import (
+	context "context"
 	logr "github.com/go-logr/logr"
 	gomock "github.com/golang/mock/gomock"
 	v1alpha1 "github.com/openshift/managed-upgrade-operator/pkg/apis/upgrade/v1alpha1"
@@ -35,9 +36,9 @@ func (m *MockClusterUpgrader) EXPECT() *MockClusterUpgraderMockRecorder {
 }
 
 // UpgradeCluster mocks base method
-func (m *MockClusterUpgrader) UpgradeCluster(arg0 *v1alpha1.UpgradeConfig, arg1 logr.Logger) (v1alpha1.UpgradePhase, *v1alpha1.UpgradeCondition, error) {
+func (m *MockClusterUpgrader) UpgradeCluster(arg0 context.Context, arg1 *v1alpha1.UpgradeConfig, arg2 logr.Logger) (v1alpha1.UpgradePhase, *v1alpha1.UpgradeCondition, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "UpgradeCluster", arg0, arg1)
+	ret := m.ctrl.Call(m, "UpgradeCluster", arg0, arg1, arg2)
 	ret0, _ := ret[0].(v1alpha1.UpgradePhase)
 	ret1, _ := ret[1].(*v1alpha1.UpgradeCondition)
 	ret2, _ := ret[2].(error)
@@ -45,7 +46,7 @@ func (m *MockClusterUpgrader) UpgradeCluster(arg0 *v1alpha1.UpgradeConfig, arg1 
 }
 
 // UpgradeCluster indicates an expected call of UpgradeCluster
-func (mr *MockClusterUpgraderMockRecorder) UpgradeCluster(arg0, arg1 interface{}) *gomock.Call {
+func (mr *MockClusterUpgraderMockRecorder) UpgradeCluster(arg0, arg1, arg2 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpgradeCluster", reflect.TypeOf((*MockClusterUpgrader)(nil).UpgradeCluster), arg0, arg1)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "UpgradeCluster", reflect.TypeOf((*MockClusterUpgrader)(nil).UpgradeCluster), arg0, arg1, arg2)
 }
