--- pkg/upgraders/aro/aro_upgrader_suite_test.go
@@ -1,13 +1,0 @@
-package aro
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-func TestAro(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "ARO Upgrader Suite")
-}
