--- pkg/upgraders/aro/aro_upgrader_verification_test.go
@@ -1,23 +1,0 @@
-package aro
-
-import (
-	"fmt"
-
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-)
-
-var _ = Describe("ARO Upgrader Verification", func() {
-	Context("When performing ARO upgrade", func() {
-		It("Checks Upgrade is Successful", func() {
-			status, err := checkUpgradeVerification()
-			Expect(err).NotTo(HaveOccurred())
-			Expect(status).To(BeTrue())
-		})
-	})
-})
-
-func checkUpgradeVerification() (bool, error) {
-	fmt.Println("Dummy check to test ARO upgrade verification")
-	return true, nil
-}
