+++ pkg/upgraders/config.go
@@ -1,4 +1,4 @@
-package osd
+package upgraders
 
 import (
 	"fmt"
@@ -8,7 +8,7 @@ import (
 	"github.com/openshift/managed-upgrade-operator/pkg/drain"
 )
 
-type osdUpgradeConfig struct {
+type upgraderConfig struct {
 	Maintenance                    maintenanceConfig                 `yaml:"maintenance"`
 	Scale                          scaleConfig                       `yaml:"scale"`
 	NodeDrain                      drain.NodeDrain                   `yaml:"nodeDrain"`
@@ -62,7 +62,7 @@ type healthCheck struct {
 	IgnoredCriticals []string `yaml:"ignoredCriticals"`
 }
 
-func (cfg *osdUpgradeConfig) IsValid() error {
+func (cfg *upgraderConfig) IsValid() error {
 	if err := cfg.Maintenance.IsValid(); err != nil {
 		return err
 	}
@@ -87,6 +87,6 @@ func (cfg *osdUpgradeConfig) IsValid() error {
 	return nil
 }
 
-func (cfg *osdUpgradeConfig) GetScaleDuration() time.Duration {
+func (cfg *upgraderConfig) GetScaleDuration() time.Duration {
 	return time.Duration(cfg.Scale.TimeOut) * time.Minute
 }
